safari-534.53-branch/LayoutTests/http/tests/navigation/redirect-cycle.html

<html>
<head>
<script type="text/javascript">
function startTest()
{
    layoutTestController.dumpAsText();
    layoutTestController.queueLoad("resources/redirect-cycle-1.pl");
}
</script>
</head>
<body onload="startTest();">
<div>Original page</div>
</body>
</html>
