LayoutTests/http/tests/navigation/redirect-cycle-expected.txt

Page 3

============== Back Forward List ==============
        http://127.0.0.1:8000/navigation/redirect-cycle.html  **nav target**
curr->  http://127.0.0.1:8000/navigation/resources/redirect-cycle-1.pl  **nav target**
===============================================
